backend/requirements.txt
pandas
scikit-learn
flask
flask-cors
transformers
torch
python-dotenv
scikit-learn
matplotlib
joblib